perf(telegram): keep setup entry on light graph
